requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
Pillow==9.0.1
transformers==4.26.0
torch==1.13.1
